Added several new make targets: make install make test make run
Broke run-in-place.  Some peer review is needed here.  It seems
very valuable to be able to run in-place, but it's very difficult
given the configure-dependent location of scripts, binaries, and config
files.

Added: rtapi.conf, hal.conf, and emc.conf, which are generated via configure
through rtapi.conf.in hal.conf.in and emc.conf.in
install.map shows the installation