bwidget has trouble dealing with "see" requests before the tree is fully laid out.  "wait_and_see" loops until the tree is viewable, and the "yview" command returns two unequal indices.  If not, it waits a variable length of time.
I tested this by selecting the top config, the last config that is
visible when the top line is visible, the next config, and the final config.
All were visible when preselected on the next run.

I also tested with few enough configs that there was no scrolling possible.
This seems to work too.

Additionally, cursoring to the top and bottom items scrolls properly.