demonstrate using request_region() / release_region(), documented as working on 2.4.x and 2.6.x kernels by     http://www.xml.com/ldd/chapter/book/ch02.html
This will probably need to be changed/reorganized or even deferred for now,
but it shows how to use the linux request_region() / release_region()
to make sure that the parport is not also being used by something else---e.g.,
the linux 'parport' driver.

When 'parport' is not loaded, or is loaded but using a different address (two parport setups not actually tested) hal_parport will work as normal, but
parport will not load at the same address until hal_parport is unloaded.

when 'parport' is loaded, starting stepper_inch will bail with a kernel
message like:
    PARPORT: ERROR: request_region(378) failed
    (make sure the kernel module 'parport' is unloaded)
though the subsequent unload of realtime fails for a reason I don't understand:
    Shutting down and cleaning up EMC...
    ERROR: Module hal_lib is in use by hal_parport
and cascading through all the other modules.  A second 'realtime stop' will
unload everything.