Moved spindle control into the motion controller. Motion currently exports the same HAL pins that iocontrol used to export. They are commanded using EMCMOT_* commands.
This is preparing work for further improvement
(like constant surface speed machining, etc).

All GUI's need to update and get spindle relevant
data from the motion stat not from IO. (e.g. emcStatus->motion.spindle.*)

Currently the only problem lies in iosh.cc, which used
to talk directly to the iocontroller, and has no idea about
motion status. For now that part is commented out, if we
decide that this is a good way to go, all IO can be part of
motion and iosh.cc can be slightly rewritten to use that.