markup fixes for manpages: Don't use \- for dashes in item names, this gives bad formatting in the HTML.
If the docstring for the component is more than one line, then put all but the
first line in the SYNOPSIS section, replacing the default lines.  This allows
a driver to document that it uses io= and not count=

Don't document the module parameter 'count=' if option count_function is
specified.

Don't document 'newinst' in the default SYNOPSIS section unless the component
is actually constructable