use rtapi_print_msg() instead of cerr
when waiting for the master, delay a variable amount of time.  this way,
if two rtapi_apps are started at the same time and a socket from a dead
rtapi_app is still around, the chance that both of them decide to convert
to master is decreased. (there's probably a race-free way to do this, but
I haven't found what it is)