a better fix for the "360 degree arc turns tiny" problem
The original problem was that the start and end points of certain 360
degree arcs would be different by a tiny amount (for example,
1.0433070866141736 vs 1.0433070866141734) when offsets were in effect.
Originally, the case for this wasn't found.  Instead, a small FUZZ
factor was introduced; circles that were smaller than FUZZ radians
were increased to be 2PI+FUZZ radians.  (this was way back in the days
before axis was merged with emc2--
http://cvs.linuxcnc.org/cvs/axis-historical/rs274/interpret.py.diff?r1=1.10;r2=1.11)

The more recent problem was in fact the exact same problem, only this
time the start and end points were on different sides of the
discontinuity in atan2() along the negative X axis.  This gave a circle
of -2PI+FUZZ radians which was changed into a circle of FUZZ radians.
The fix just made (rev 1.3) adds 2PI repeatedly until the circle is
bigger than FUZZ radians.

This change gets rid of FUZZ and adding of 2PI more than once.  It
attacks the root cause of the start and end points being different: The
addition and then subtraction of coordinate offsets from the arc
starting point.  In floating-point math,  (a+b)-b is not always equal to
a.  By reorganizing the code, the whole arc is done in offset
coordinates, which means not subtracting the X offset from the "prior X
coordinate" (and Y and Z) and it's possible to simply rely on exact
equality of start- and end-points like emccanon.cc does.