started changing jogging NML commands (EMC_JOG_CONT, EMC_JOG_INCR, EMC_JOG_ABS and EMC_JOG_STOP)
The original version of this commit added to halui.cc a function called
sendJogIncr(), with an "r".  But after the rebase of joints_axes3 to
joints_axes4, the parent version of halui.cc in the new starting place
in master already has a function called sendJogInc(), without the "r".

joints_axes4 changed this commit to follow the naming convention in
master.