extend the naive cam detector to arcs.
When the maximum deviation of an arc from a straight line is less than the
G64 P- tolerance, break the arc into two lines (from start of arc to midpoint,
and from midpoint to end).  those lines are then subject to the existing naive
cam algorithm which may eliminate or join successive lines that are within
the tolerance.  Thus, line-arc, arc-arc, and arc-line cases can now benefit
from this algorithm, in addition to the classic line-line case.

In BigJohnT's situation with output from sheetcam this increased his cam
contouring speed from 150IPM or less to >400IPM.