restore soft limits box for lathes
the old code got limits like [x, 0, z] while the new code gets limits like [x, +-1e99, z].
This is apparently an implementation detail in task when initializing limits
for axes that are not used.  So let's treat any number bigger in absolute value than 1e30
as 0, so that lathe limits get drawn in the plane y=0 which is the plane of the
screen, instead of outside the view volume.