cooperate with linux parport_pc driver if it is loaded
if parport_pc is loaded, use the linux parport layer to request access to the
underlying port, instead of going through rtapi_request region.  as long as
the port is registered and claimed, linux should prevent other device drivers
from using the same resource.

in addition, when parport_pc is loaded, I/O addresses below 16 are treated
as port indexes.  e.g.,
    loadrt hal_parport cfg="0"
will use the address Linux has detected for parport0.

if parport_pc is not loaded, or the requested port is not known by parport_pc,
then the old rtapi_request_region code is still used.  This means that systems
which block loading of parport_pc will also continue to work.