allow all pins, including special-purpose output pins, to be read as inputs
It's possible to sample the value on hostmot2 physical I/O pins even
when they are used as outputs--either gpio or special-purpose modules
like stepgen.

Use 1 (most useful): For an open-drain output, the pin can be driven
low by an external source.  This allows a sort of one-wire handshake.

Use 2: You can read the output of a special purpose module.
Reading the output of a special-purpose module probably has limited
usefulness--since it samples at a very low rate, you can't see the true
step+direction waveform, just like sampling a software stepgen's step
pin in the servo-thread of a system where it runs on the base-thread
gives misleading results.

But it might be useful in a few special circumstances; for instance, by
properly setting the step timings it might be possible to guarantee
catching every output transition, and this could be used in some kind of
testing.

Use 3 (least useful): For the sake of consistency, a gpio "in" pin now
tracks the corresponding "out" pin, for push-pull outputs.