Add INI options code / mesa work
Add code to print INI options in HAL file
little more work on the signal blocking on
mesa page. Mesa page configure automatically
if a config is reloaded or waits for user to
choose components etc if not loaded. Will
not let user pass mesa page if component type
not configured.