hm2 stepgen bugfix for float-precision logspew problem
The x86 has 80-bit floating-point registers, but doubles in RAM are only
64 bits wide.  Storing a register to RAM results in truncation and loss
of precision.

The driver has a bug where a value in RAM (stepgen.maxvel param) is
compared to an arithmetic result (physical maxvel) computed from inputs
in RAM (stepgen timing params).  If maxvel is too high, it prints a
warning and sets maxvel to the physical max.  However, this assignment
is a store from a register to RAM, so it loses precision, and the print
gets hit again next servo period.

Jeff's fix (implemented in this patch) is to force the computed value
down to 64 bits after the computation, before doing the comparison.