Use a high-resolution probe result
Connect a "latched position" input, such as encoder.#.position-latched,
to the new axis.#.motor-pos-probed pins to get single-count precision
probe results regardless of probe velocity or servo cycle time.

At the moment only the software encoder counter supports the
high-resolution latched position, but it is anticipated that the
software step generator and the hostmot2 encoder counter will be
adaptable to this system.

For existing drivers, connect the axis.#.motor-pos-probed to the same
thing as motor-pos-fb for drivers which don't support high-resolution
probe results.