Stop the pointed-at button from looking different
this visual behavior doesn't make sense on a touchscreen system.
The two different rc snippets are for the Hardy and Dapper versions
of gtk, respectively.  The Hardy version should automatically adapt to
themes with different base colors because the gtk_color_scheme setting
is overridden by the user's settings, but I haven't tested this.