preserve fms values for nonrandom tc, and more
To get compatibility with existing tooltables [on nonrandom machines],
we need to think of the first column as the tool number.  The second
column is unused but we can preserve it without too much trouble.
I bet this is the least surprising behavior.

For random, put the tool first as well, for consistency.

Also, improve the error messages (though they still bogusly go
to stdout.)