cut out extra extras directories
As far as I can tell, there were two main things that differentiated
the extras directories, besides unintentional differences that
crept in.  First, in 6.06 we had some udev stuff that should
have been in the rtai package.  Second, we had the parport-disabling
stuff that we didn't want for sim packages.

emc 2.4 won't support ubuntu 6.06, so there's no need for that hack,
and the parport won't be disabled by default so it's safe to install
the do-nothing file /etc/modprobe.d/emc2.

The ability to have a different extras directory is retained, but
starting now all the distributions will get the same extras directory.