better encoder velocity estimation when reversing
The trick i was missing was that you need to keep track of the previous
movement's (not the previous servo-period's) direction, independent of
the reported velocity.

Thanks to Peter Wallace and Chris Radek for the bug report, and to Chris
Radek for testing a couple of iterations of this patch.