accept IDROM Type 2 and Type 3
The only difference is that type 2 has a "primary gtag=0" sentinel at
the end of the Pin Descriptor array, and type 3 does not.

Both IDROM Types have fields in the IDROM header that specify the length
of the PD array (IDROM.IOPorts * IDROM.PortWidth), so the sentinel is
redundant and wasteful.  The current version of the driver uses the
expected length of the array for both IDROM Types.