Make the hm2->pin array dynamically sized
The IDROM says how many IO pins the board has, so we dynamically allocate
the pin array to be that big (and free it when we're done with it).

Instead of reading Pin Descriptors until we find a pin where the primary
gtag is 0, we now just read as many PDs as the IDROM header said to expect.