Show status of mdi words in a less confusing way?
Previously 'greyed out' meant the word won't be
sent when the command is issued.  this is not the
standard use of 'greyed out' which typically means
the user can't interact with the widget.

Two columns for 'selected' and 'not selected' is
a somewhat standard gui pattern so maybe this is
less bizarre?