paper over a memory leak
each call to open_file_guts leaks the old GLCanon object, o.g.
the canon object keeps references to information like the endpoints
of each line in the program, which can add up to a lot of memory.

it would be ideal to not leak the GLCanon object, but in the meantime
calling its new .clear() method can reduce the memory from megabytes
per reload to kilobytes per reload, which greatly reduces the impact
of the bug.