Fix 'checking for ANSI C header files... no' with --enable-simulator
It appears that if 'egrep' hasn't been searched for earlier, it is
searched for at the site of the first AC_CHECK_HEADERS (indirectly via
AC_EGREP_CPP), but this site happened to be conditional on RTS != sim.
By moving these tests to an unconditional site earlier in configure.in,
egrep is detected and ANSI C header files are found.