Work around a numarray slicing bug (SF#2947390)
On some platforms, numarray slicing seems to be broken.  On these platforms,
a slice like [4:5] gives 5 items instead of 1 item:
    # Python 2.6.4, numarray 1.5.2, numpy 1.3.0, amd64 (Ubuntu 9.10)
    >>> l = [0]*10; z1 = numarray.zeros((10,)); z2 = numpy.zeros((10,))
    >>> len(l[4:5]), len(z1[4:5]), len(z2[4:5])
    (1, 5, 1)

On other platforms, numarray slicing behaves as expected:
    # Python 2.4.3, numarray 1.5.1, no numpy, i386 (Ubuntu 6.06)
    >>> l = [0]*10; z1 = numarray.zeros((10,))
    >>> len(l[4:5]), len(z1[4:5])
    (1, 1)

The same problem applies to slicing of 2D arrays, which is one of the
primary activities of image-to-gcode.

The reformulation of
    arr[l:l+w, t:t+h]
into
    arr[l:,t:][:w,:h]
gives a slice of the desired items even on systems with broken numarray.