Fix RFL on line that does not specify all axis words
CANON_UPDATE_END_POINT is not enough to get the actual start point of the
first non-skipped line to the interpreter.  If the following line was
'G0Z0' it would go to Z0 and the X and Y location of the previous (skipped)
move.  The same was true of an M6 toolchange when it uses the quill-up
location