fix debian package version numbers on the buildbot
"git describe" shows the most recent (timewise) tag in the history.
When we merge a branch that has tags into another branch, for example
merging 2.4 into master, the tags from both branches get searched by
git-describe.

Since we use tags to make the version numbers of our debian packages,
this can lead to incorrect debian versions.  Debs in the master repos are
always builds of the master branch, but they sometimes would incorrectly
claim to be 2.4 debs....  Debs in the 2.4 repos are always builds of the
2.4 branch, and they always have correct 2.4 version numbers (since we
haven't merged any branch with tags into 2.4 yet).

This commit fixes the way we run git-describe so it only searches for
tags appropriate for the current branch.