Support kernels 2.6.29 and newer (ie Lucid Lynx)
Before 2.6.29, request_firmware() gave the driver the firmware in a
writable buffer.  The hostmot2 driver needs to reverse the bits of each
byte in the .BIT file before sending it to the FPGA.  Up until this commit
the driver re-wrote the buffer in place.  Kernels 2.6.29 and newer give
the driver the firmware in a read-only buffer, so the driver page-faults
when it tries to write to the buffer.

This commit makes it so it doesnt try to write to the firmware buffer,
instead it does the rewriting on the fly as it's sending the data to
the FPGA.

Here's the linux kernel commit that made struct firmware's data pointer
read-only:

    <http://git.kernel.org/?p=linux/kernel/git/torvalds/linux-2.6.git;a=commit;h=6e03a201bbe8137487f340d26aa662110e324b20>