be more verbose when testing
Up until now there have not been any tests that got far enough to read
the Pin Descriptor array or the Module Descriptor array, so there was
no need to dump what we found there.  But I'm about to add a test with
a valid PD array, and I want to see what it is.