fix linking when compiling for installation
When we're compiling for installation (manual installation or via the
debs), we want our executables and libraries to be linked against
our not-yet-installed libs, and we don't want our build dir in the
executables' rpath.  "rpath-link" does exactly this.

When we're compiling for run-in-place, we want our build dir in the rpath
(which also lets the linker find it), so nothing changes in this case.