fix a stupid bug in my previous commit
LIB_DIR is where we're planning to *install* the libs, not where we're
currently staging them.  We stage them always in "../lib".

(I didn't notice this because I stupidly had emc2.deb installed, so
LIB_DIR *did* indeed have the libs in it...  Third time's the charm
I guess/hope.)