reduce default number of segments
This cuts the default maximum number of segments of an arc in
half, from 128 to 64 for a half circle.  The maximum deviation
from a radius-1 arc to a straight-line approximation where the
included angle of each segment is theta is 1-cos(theta/2).  The
old value gives a deviation of 0.0075% and the new value gives
a deviation of 0.03%, or 1 mil deviation on a circle with a
radius of 3 inches.

To use the old value, add [DISPLAY]SUBDIVISION=128 to your
infile.