Fix some minor linking issues
Many of our binaries are linked by g++, which links in a bunch of C++
libraries by default.  But several of those binaries didn't need or want
the C++ libraries, they're just superfluous junk linked in.  This commit
changes the makefiles for those binaries to link using gcc instead of g++,
which fixes this problem.

Also, halrmt was linked against readline, but never used it, so I took
that out.