don't prevent window from being shrunk
this code interacts badly with the paned widget, because after
the sash is moved, .pane's reqheight becomes its actual height.
by doing it only once at startup (instead of anytime the root
window is reconfigured), the effect is avoided.

I was unable to determine what the "+4" fudge factor was for; the old
commit in the AXIS CVS just says "change wm minsize recalculation"!