on tk8.5, make top pane enlarge by default
Before the addition of the pane, the top part would enlarge and the
program text would stay fixed-size when the window was enlarged.  After
the addition of this patch, the text enlarges.  Tk8.4 (Ubuntu hardy)
offers no control over this, but Tk8.5 (Ubuntu karmic) does through
"paneconfigure -stretch".  When this feature is available, restore
the old behavior of enlarging the preview area by default, keeping
the program text fixed-size.