don't allow the pane to shrink too much
Chris identified a problem when you: maximize the window, then make
the text area small, then unmaximize the window.  The text area would
become completely hidden.  This makes sure both the top and bottom
areas of the pane retain at least their desired size.  It reduces
flexibility of the pane, but prevents the program text from appearing
to be "lost"