make sure the selected line stays visible
when the pane or window is resized, the number of lines of program
text that are visible can change.  Scroll back to the running or selected
line in this case.

Unfortunately, this behavior happens even if the user has explicitly
scrolled away from that line.  However, this is probably a less likely
circumstance than the one that this is intended to correct.