make a certain "full circle" arc appear
with specific offsets in effect (G10 L2 P1 X1.2 Y3.4 Z-3.1 R33) the yz
circle in 3dtest.ngc didn't appear.  The arc is intended to be a full
circle but it ended up with theta1=-2.35619449019234439
theta2=-2.35619449019234484, indicating a clockwise arc with
an included angle of about 4e-16 radians (.5**51).

this change also incorporates a workaround for another "full circle" arc,
the one where the atan2 calls give +-PI (the range of atan2 is +-PI
inclusive; these occur for atan2(y,+-0), with the sign of 0 being
important).  In this case, the first addition of 2*M_PI to theta2
can turn a negative one turn circle into a zero turn circle; two
additions are required to turn it into the desired positive one
turn circle.  This specific misbehavior hasn't been seen yet, though,
only anticipated.