Add a relative tolerance for arcs
Previously, IJK arcs were rejected based on the absolute radius error alone.
This is problematic when the numbers come from calculations where at least
one input is a non-exact constant given as a decimal number.

Adding a relative error check (arbitrarily .1%) and rejecting an arc only when
it fails both the absolute and relative error checks, or when it fails a much
larger absolute error check allows large arcs with relatively small errors to
be accepted.