Approximately fix overblending of arcs.
Especially tangent arcs would have a blend that did not respect
the tolerance setting.  The blending algorithm assumes that if
there is perfect tangency the blend can be as large as needed,
since there will be no deviation from the path.  This is true for
line-line blends but not line-arc.

This change fudges the start-direction vector in the arc case
toward the direction of the centripetal acceleration.  The
tolerance calculation is still done as if the motion that is
starting is linear, but now in a more representative direction.

This is only an approximate fix but testing done my micges and me
shows that it looks OK.