Fix hm2 stepgen.enable=0 behavior
This commit fixes SF#2985881 "stepgen bug in hostmot2".

This commit changes the hm2 stepgen.enable behavior.  When .enable is
true it behaves like before, but now when .enable is False it behaves
like this instead:

    No steps are generated (if the stepgen was moving at the time enable
    went false it stops immediately, without obeying the maxaccel limit).

    .position-fb starts tracking .position-cmd, and .counts changes to
    match (even though no motion is taking place).

    .velocity-fb goes immediately to 0.

This makes it so that when .enable later becomes true again, no abrupt
motion takes place.