parport: Don't believe linux parport feature bits
Neil B reports a ppmc system which works properly with emc 2.3.5
but doesn't work in 2.4 which fails with the message "parport0 does not
support mode 4" (EPP).  Instead of failing when Linux doesn't detect
the desired mode, continue anyway after printing a stern message.

In the case of EPP, all the drivers immediately try to communicate with
the attached device so actual lack of EPP will still end up causing a
registration error.