axis: avoid leaving lots of zombie processes
When creating a process with spawnvp(P_NOWAIT) it's necessary to
subsequently waitpid() the child process.  When this isn't done,
it appears as a zombie process until its parent exits.

Instead of adding the necessary machinery to axis to track the
PIDs returned by spawnvp, use tcl's [exec ... &] which has its own
(slightly odd) machinery for reaping child processes.  In normal
circumstances (not launching multiple editors at once) this means
there will be a small number of zombies, not an ever-growing number.