tcl packages for use by external tcl programs
Locate emc.so, hal.so, emc.tcl and a new pkgIndex.tcl file in a directory that
is a subdir of a directory included in the tcl auto_path list as configured for
the tcl/tk packages in the ubuntu distribution.

For a conventional build with --prefix=/usr, the chosen parent dir is:
   /usr/lib/tcltk
and the shared library files will be located in the subdir:
   /usr/lib/tcltk/emc2

With these changes, aan external tcl program should be able to say:
   package require Emc 1.0
and/or
   package require Hal 1.0

For this normal case, no extra configuration is required.

For an installed emc built with --prefix=/otherlocation, the using
application must modify the tcl auto_path list:
  e.g., lappend auto_path /otherlocation/lib/tcltk
or set the TCLLIBPATH appropriately
  e.g., export TCLLIBPATH=/otherlocation/lib/tcltk

For a run-in-place build of emc, '. scripts/emc-environment' to set
TCLLIBPATH appropriately.