provide a function to calculate nurbs tangent
NURBS_FEED uses an odd method to compute the nurbs tangent.
This method is still ad-hoc but at least it's encapsulated
and should have better numeric accuracy than the original
method which involved widely-separated samples of the NURBS
and trig.

Later if direct calculation of nurbs tangent is added, it
will fit nicely within the API established here.