Save state of G92 correctly across restarts
Previously if you had a nonzero G92 offset that was unapplied with
G92.2 and you restarted emc, you would get the G92 offset applied.

This change records in the var file whether the G92 offset is actually
applied.

This also fixes AXIS's preview when there is a saved but unapplied
offset.