classicladder: modbus: Fix crash if LgtResponse==1
When LgtResponse was 1, CRC16 would be passed a length argument
of (unsigned short)-1, which would cause it to go past the storage
allocated for Response.

The modbus crc is designed so that after reading a packet and its
crc, the new crc is 0.  Using this property, calculate the CRC on
LgtResponse bytes and check that result, instead of checking the
CRC on all-but-two bytes against the last two bytes.  This eliminates
the need for a special case on a 1-byte response.