Fix load failure on -200s with parport_pc loaded
For reasons that are not yet understood, when hm2_7i43 cooperates
with parport_pc, hm2_7i43_reset fails to reset the 7i43-200 card,
and then misidentifies the card as 7i43-400 which makes the driver
fail to load.

While the underlying cause is not yet understood, the additional
attempt to request the FPGA self-reset fixed the problem on the
test system.