fix a stepgen bug
Before this commit, the stepgen would sometimes fail to slow down all the
way when decelerating to a stop after completing a move in the negative
direction.  Even worse, it would then fail to notice that it was still
moving (vvvveeeeeerrrrryyyy slowly) in the negative direction.  It would
creep in the negative direction, at something like one step per minute,
but the .position-fb and .counts pins would not report the movement.
The steps were just lost.

This commit fixes an integer underflow bug, by casting the variables up
from u32 to s64 before doing the arithmetic.