fix tan; provide atan
The x86/x87 implementation of tan in terms of fptan was wrong, because
fptan puts *two* values on the stack: tan(theta) and 1.0.  This is
apparently for some sort of symmetry with fpatan.

The implementation of atan was missing altogether.  It is no surprise
that atan can be implemented in terms of atan2, just as the other inverse
trig functions are.