hm2_pci: indicate failed registration to userspace
Before this change, hm2_pci would still return success even if
no boards were registered, or if some boards failed registration.
This obscures the real problem, usually making the failed command
be a halcmd 'net' or 'setp' command for a hm2_pci pin or parameter.

Now, if zero boards are detected, or if at least one board encountered
a failure to register, rtapi_app_main will return a negative errno value
to indicate the problem, and halcmd will indicate the 'loadrt' line
as the cause of the problem.