axis: fix keys while choosing jog increment (SF#3028197)
The combobox wants to intercept up, down, pgup, pgdn keys for its
own purposes while the dropdown is open.  However, the combobox did
not intercept the corresponding KeyRelease events.  This interacted
badly with the AXIS main window, which uses key release events to
update the "current axis".  By intercepting the release events in the
same conditions as the press events are intercepted, this problem
is avoided.