Fix g98/g99 to match fanuc retract planes behavior
This makes it possible to switch retraction between "retract" (R)
and "initial" planes with g98/g99 during a series of cycles.

Previously, the "initial" plane was lost (set to match R plane)
when the first g99 cycle executed.

RS274NGC3 says g98 retract is "to the position that axis was in
just before the canned cycle started ..." and now this is changed
to mean "to the position that axis was in just before this series
of one or more contiguous canned cycles was started"

The "initial" (g98) plane is reset any time cycle motion mode is
abandoned, whether explicitly (g80) or implicitly (any motion code
that is not a cycle).  N.B. switching among cycle modes (say g81
to g83) does NOT reset the "initial" plane.

This may affect the operation of some existing ngc programs, but
in a safe way: the change can only cause more retraction, never less.