Put commanded spindle speed on a hal pin
The commanded spindle speed (before interference of spindle override) is
useful when selecting the spindle gear.  Attempts to calculate this
value by dividing the command by the override are flawed because the
override (from halui) and speed (from task) do not update in synch, and
because of the case where spindle override is set to 0%.

When constant surface speed is in effect, the "commanded" speed is set
to the maximum speed.  While this means the name "commanded" is not
quite accurate, it will give the right information to a gear-shifting
system to choose an appropriate speed for the operation.

Only a "revolutions per second" (RPS) pin is provided for this new
output.  If it turns out to be difficult to adapt gear changers to
accept this instead of an RPM output, an RPM output could be added later
in the obvious way.