update debian/configure script for Debian Lenny
Debian Lenny (5.0) does not have a /etc/lsb-release file, but it does
have the lsb_release program, which is a command-line interface for
fetching the same information as that file has on Ubuntu.  Ubuntu Hardy
through Lucid (and probably others) have the lsb_release program too.
This commit makes the debian/configure script prefer the lsb_release
program over the /etc/lsb-release file as the source of this information.

This commit also adds specific EXTRA_BUILD info for Lenny, since the
default EXTRA_BUILD specifies lyx-qt, which is not available on Lenny.