Fix doc build error introduced at 30341a87
Apparently index entries which contain underscores don't
build on Ubuntu 8.04 unless a backslash is added before
the underscore.  This is related to the way LaTeX normally
treats _ as denoting a subscript in equations.

This unfortunately adds spurious backslashes in the HTML index,
but we'll have to live with that at the moment.