tkemc: fix radiobutton appearance
A difference between tk8.4 and tk8.5 radiobuttons meant that
on Ubuntu 10.04 it was not possible to see whether a radiobutton
was selected or not selected.

After this change, the background of the checkbutton's indicator is
blue and its indicator dot is white.  This seems most in keeping with
the tkemc white-on-blue style.