On random tc, correctly skip invalid tool table entries
This fixes an error case on random tc when a tool table entry
described a pocket that was out of range.  This printed an error
message but then if there were subsequent valid words on the line,
the entry was incorrectly marked as 'valid' again, and the invalid
pocket number caused the array to be written out of bounds.

To reproduce, on a random tc config, use a tool entry such as
T1 P99999 Z1.0

With this change, the invalid entry is skipped when reading, so
when the tool table later gets rewritten it's gone.  I think this
was the original intent.