tooledit: fix bug with 0 at start of comment
In d8583c4, a bug was introduced that made a comment that started with 0
be treated by a number.  This meant that entering a comment like '0x0'
woud lead to an error (and trashed tool table) on save.  A comment like
'0.01234000' would be turned into '0.01234'.

Change this behavior so that the comment field is never treated like a
number.