rs274: use locale-independent number parsing
Use C++ string streams to perform "C locale" conversions regardless
of the global locale settings.  C++ solves the global nature of
locales by requiring the stream to be "imbued" with a locale before
any locale-specific settings affect the stream operations.  If a
stream is not imbued, operations on it (such as >> to parse into a
double) use C-locale-like settings.

After this change, it is possible to modify the standalone
interpreter to call setlocale(LC_ALL, ""), set
LC_NUMERIC=de_DE.UTF-8, and still parse normal gcode like "G0 X0.0"
(it also will still reject code like "G0 X0,0")

strtod always honors the current LC_NUMERIC locale, so when
LC_NUMERIC indicates a locale where the radix character is "," (or
any other value besides "."), the old formulation in terms of
strtod() behaved in an undesirable fashion, leading to the spurious
error "Bad character '.' used".

In the past we have handled this by the (undocumented) requirement
that the interpreter only be called with LC_NUMERIC=C.  (for
instance, emctaskmain only sets the LC_MESSAGES and LC_CTYPE
categories, rather than setting LC_ALL).

Another suggested solution is for the library to change LC_NUMERIC
when entering and exiting its APIs.  However, because locales are
not thread-specific, this strategy would interfere with a
hypothetical program that has multiple threads and desires
LC_NUMERIC be in the user's locale--in thread 1 the interpreter
temporarily sets LC_NUMERIC to C, while in thread 2 the gui does a
number parsing or formatting operation which expects LC_NUMERIC to
be the user's choice such as de_DE.UTF-8. (on the other hand, before
the present solution, the application could put locking and
setlocale calls around each call into the interpreter, to ensure at
the application level that each portion of the software sees the
locale that it requires to function properly)