halui: reduce duplicated structure definitions
it's possible to use C++ to declare both structures, so that
the number of places to change when one field is added is reduced.

This also has the effect of removing the volatile qualifiers from
the members of local_halui_str.  These volatile qualifiers are unneeded,
because (unlike the members of halui_str) halui has total control over
when members of local_halui_str are modified.