Fix XY TLO in rotated systems, and add G10 L11
Tool length offsets don't rotate with the XY coordinate
system.  This was inconsistent and caused settings->current
to become wrong after using G43.

Also I added G10 L11 which is just like switching to G59.3,
then using G10 L10, then switching back to the original system.

This is to allow a fixture that does not follow the workpiece
(for instance a 123 block on the table) to be used for touching
off tools.  Just set G59.3's origin at the fixture.

GUIs need support for somehow optionally generating the G10 L11
code instead of G10 L10 when setting a tool.