fix key_t vs char* confusion
At the innermost and outermost layers, rcs semaphore keys are
small numeric values (e.g., one such key commonly used in emc is 0x3ea).
Somehow in the middle layer the types used for passing keys were char*,
which led to "FIX-ME fsck'd ugh" and on 64-bit systems a warning about
a cast to a narrower type.

Make the argument type key_t in all layers, which reduces the number
of casts and WTFs overall.