build: make kbuild aware of rtai symbols
Some time in the past, we had rtai packages which lacked Module.symvers.
This meant the only way to avoid the "undefined symbol" warnings for rtai
symbols during the emc2 build was to silence them with the modsilent
script.

However, as our Ubuntu 10.04 rtai packages ship this file, we can tell
kbuild about it, and get only real undefined symbol warnings.

Unfortunately we still can't do away with the modsilent wrapper, due to
the way we deliberately multiply define the kinematics functions in each
kinematics module.