config: fix --as-needed vs sincos build problem
LDFLAGS=-Wl,--as-needed ./configure would fail to detect sincos.  Worse,
under some gcc versions the replacement sincos would be optimized by gcc
into a call to the function sincos, which led to a recursive crash on
that platform.

By putting the math library in LIBS instead of LDFLAGS, it comes after
the source file.  With --as-needed AND when -lm was specified before the
source file, the linker legitimately could omit the math library,
leading to the "undefined symbol 'sincos'" error that configure detected
in the test.