hm2_pci: fix unload after failed registration
A bug was introduced at 44e7da328: after a failed registration (e.g.,
due to a bad firmware path) the hm2_pci kernel module could not be
unloaded because it still had a PCI device registration.

(in rtai, at least), rtapi_app_exit isn't called called when
rtapi_app_main returns with an error code.  Therefore it's necessary
to pci_unregister_driver in the error paths that follow the registration.