Fix current tool information after reloading table
The earlier "fix" c57f2d2 was busted from the start.  On a nonrandom
system, realod_tool_number() always called load_tool(0) which unloaded
the current tool!

Machines with random toolchangers never had this class of bug.  For
nonrandom systems, unless you start with pocket 1 (not 0) when searching
for the pocket that contains the desired tool you will always unload the
loaded tool instead of updating the status buffer with new tool
information, because you'll *always* find the current tool in pocket 0.