interp: reject ABC in canned cycles
In canned cycles, ABC motions are not permitted.  However, this error
sometimes passed silently.  For instance, the following wouldn't
provoke the error but also would not perform the desired motion:
    g81 x0 y0 z-1 r0 f12
    a90
this happened because g_modes[1] is not set until later.  At this point,
motion_to_be should be relied on to check the type of the current motion.