build: tame the stack size warnings
gcc 4.4 introduced -Wframe-larger-than=N to produce warnings when the
stack frame of a function is larger than a certain size.  The linux
kernel, which uses 4kB or 8kB stacks, sets this warning threshold at
1024 bytes.

We hit this warning in a number of places, but because hal uses a stack
size of 16kB, the default warning threshold is too conservative a choice
for us.  Setting the limit to 2560 bytes silences all the warnings
currently in the software, including the largest stack frame of 2256
bytes in genserikins.c:kinematicsInverse().  We have reports from
alex_joni that genserkins works in the field, so there's every reason to
think that the new warning limit of 2560 bytes is reasonable.