silence warnings in rtapi_shm.h
on lucid, there are compile warnings that originate in rtai_shm.h.  We
do not care about these warnings, so silence them by telling gcc that
rtai_shm.h is a system header.

The warnings are of the form
    rtai_lxrt.h:615: warning: 'rtai_lxrt' is static but used in inline
                     function 'rt_get_adr' which is not static
which does indicate a problem, but one which does not appear to affect
us and which must be fixed by the developers of rtai.  For more
about this warning and its motivation, see e.g.
http://coding.derkeiler.com/Archive/C_CPP/comp.lang.c/2008-10/msg01436.html