pid: avoid a 1-cycle glitch at encoder index
just like for FF1, the new method for computing the D term (when the
*exactly one* of command or feedback has a high-quality velocity estimate)
is affected by the position jump.  To avoid this, skip computing the
low-quality velocity estimate on the cycle where the index-enable bit
falls.

Before the high-velocity estimate code was added, the jump didn't matter
in the computation of the D term because an identical jump was present
in both the command and feedback halves of the equation.