fix compile on kernels 2.6.24 and older
In 2.6.25, the items in 'struct pt_regs' had their names unified on x86
and amd64, so that the instruction pointer was 'ip' on both systems (and
actually on a range of other architectures as well).  In the bad old
days, struct pt_regs had differently named fields on the two
architectures in question, eip and rip.  Accomodate the different
possibilities.