ppmc hm2_7i43 pluto: use specified base_hi as fallback
With specific motherboards (e.g., D510MO), Linux detects the parallel
port but does not detect a port_hi.  This makes the hal epp drivers fail
to work.  When using a linux-detected parport without a linux-detected
high address, use the same method to choose a port_hi address we already
use for non-linux-detected parports:
 * -1: force no port_hi
 *  0: use port+0x400
 * >0: use specified value
The port will be registered with request_region, so it will not
interfere with any other linux-detected device--in case of a conflict,
the hal_parport_get will return success but port_hi will be 0.  as now,
the epp drivers will fail to load.

There's a small possibility that a non-detected device will exist at
the specified port_hi, but this is no worse than using the port without
cooperating with parport_pc.