Fix rare problem with freqgen output getting stuck "on"
If make_pulses hadn't run before update runs the first time, the
initial timing calculations would be wrong (unless make_pulses
happened to be in a thread with the default period of 50000 ns)
The period is now checked for changes, and a recalculation of the
PWM parameters is triggered if it changes.

This also allows you to change the thread that make_pulses is run from
while the system is running, and it will compensate for the change.