fix a never-tripped NULL pointer dereference
This commit fixes a bug that never bit.

The probe_new() function pops up a window that lets the user select
what to show in the meter, and it takes a string which is used as the
title of that window.  probe_new() tried to protect against the caller
passing in NULL for that title string, but the check was reversed:
if a non-NULL string was passed in, the probe_new() function would
replace the string with a default.  But if a NULL string was passed in,
the check would incorrectly accept the NULL string as the window title,
and shortly thereafter segfault.

This bug never bothered anyone, because there is only one place that calls
probe_new(), and it passed in a non-NULL string (which was replaced by
the default window title in the bogus NULL check).

This commit fixes probe_new() function to check for NULL correctly, and
fixes the caller to pass in the window title that was actually being used
(the default in the broken NULL check).