skip the hm2_idrom test on 64-bit hardy
This test loads the hm2_test driver, which does things like print out
long lists of hm2 pin descriptors into the kernel log (dmesg).  Then in
userspace the test greps the dmesg output for special messages from the
hm2_test driver, to verify that it did the right thing.

This works great on on hardy-rtai-32 and lucid-rtai-32.

But for some reason 64-bit Hardy (at least in KVM) doesn't put all the
kernel messages into the dmesg buffer, which causes this test to fail.

Here are dmesg output of runtests, on 32-bit and 64-bit hardy:

http://emc2-buildbot.colorado.edu/buildbot-admin/builders/hardy-i386-trunkish-realtime-rip/builds/21/steps/get-dmesg/logs/stdio
http://emc2-buildbot.colorado.edu/buildbot-admin/builders/hardy-amd64-trunkish-realtime-rip/builds/6/steps/get-dmesg/logs/stdio

Search for "test pattern 12", and note that on 32-bit, all the pin
descriptors are included in the dmesg output.  On 64-bit, however,
the output is trunkated near the beginning.

It's not a LOG_BUF_SHIFT kernel config issue, since it's the middle
(rather than the top) of the dmesg that's missing.  I don't know why
it's behaving this way, so i'm disabling this test on 64-bit Hardy.