fix a segfault
Interp::doLog() opens its log file before writing to it, but it would
write to the file handle even if the open failed.

The funny thing is it checked for failure-to-open so it could write
an error message, but then it went ahead and wrote to the NULL log
file anyway.