Fix spurious cutter comp error with tiny arcs in metric
http://www.linuxcnc.org/component/option,com_kunena/Itemid,20/func,view/catid,38/id,5850/lang,english/#5850

This fix is the safest possible one for late in v2.4: it changes
no logic, it just adapts the existing fudge factor to be correct
in mm too.

The error is only a few microns, but I wonder if we can improve
the math (in find_turn maybe?) to fix it earlier.