milltask: don't busywait motion
There is a spike in CPU activity when filling the realtime motion queue.
'perf' indicates that this is due to the busy-wait in
usrmotWriteEmcmotCommand (90+% of time spent in the kernel portion of
gettimeofday).  Adding a modest (25us) sleep after each failed poll
pushes this down to 4% of the time and gets rid of the CPU usage spike.
It doesn't appear to lower the queue fill rate measurably, at least as
long as the servo period is long compared to 25us.